Fix double payment issue
If you configure payments to be too frequent, a call to the payment
daemon may still be executing while another payment loop begins. This
could lead to the second (or more) loop fetching the some of the same
share information as the first and making extra payments. This is
evident when balances go negative.

This commit adds the ability to set a timeout for calls. The default
timeout is 60 seconds but is overriden by another change to
paymentProcess.js in z-nomp that sets the timeout to half of the payment
processing time alotted by configs. This ensures some of the more time
expensive calls don't stick around long enough, leading to double (or
more) payments.