ZelCash nodes (#34)
This PR adds payouts to all three tier of zelnodes (zelcash masternodes). Information is taken directly from getblocktemplate rpc call. If there isn't an enabled zelnode, the information is undefined and payout goes to miners.

Payouts are paid even if they are not yet enforced by the spork. Enforcement of payouts will be activated through spork after 21st of Feb.

payZelNodes can either be set to true or to a timestamp (in seconds!) of when expected enforcement of zelnode rewards by spork activation is about to happen for automatic transition. When false or unspecified zelnode payouts wont be paid.

The expected time of zelnode payouts is around 25th of Feb (TBD) so pool operators update their backend and specify "payZelNodes": 1551103200 in their zel.json file and once that time passes zelnode payouts will be paid. This still checks if the conditions for payouts are present - so if daemon is not specifying a zelnode payout address with a value, the payout will go to miner and not to zelnode even if payZelNodes is active.