Bugfix: Use 'getblocksubsidy curHeight + 1' to determine coinbase tx vout amounts
Fixes a bug where invalid blocks are generated when a coin reaches a
SubsidyHalvingInterval block.

node-stratum-pool uses the 'getblocksubsidy' RPC call without the optional
'height' argument, which returns the block subsidy reward at the current
block height. Instead 'height' with the block height of the next block
needs to be passed to 'getblocksubsidy'.

As the values from 'getblocksubsidy' are used to set the amounts of coinbase
tx vouts, upon reaching the halving block height boundry, node-stratum-pool would
generate a block that doesn't take the halved block rewards into account, which
gets rejected by the coin daemon.